executor: Move all runtime execution tests into tests file.
