+++ core/executor/src/lib.rs
@@ -38,6 +38,8 @@ mod sandbox;
 mod allocator;
 mod host_interface;
 mod wasm_runtime;
+#[cfg(test)]
+mod integration_tests;
 
 pub mod error;
 pub use wasmi;
